Enhance dictionaries structure
