+++ packages/core/base/src/utils/i18n/dictionaries/Klingon.ts
@@ -1,0 +1,16 @@
+const Klingon = {
+    crossmintPayButtonService: {
+        CONNECTING: "yImej...",
+        BUY_WITH_ETH: "ETH vItlhutlh",
+        BUY_WITH_SOL: "SOL vItlhutlh",
+        BUY_WITH_CREDIT_CARD: "QelI'qam vItlhutlh",
+    },
+    crossmintStatusButtonService: {
+        INVALID: "mIw clientId",
+        WAITING_SUBMISSION: "ghItlh Crossmint vImej tlhej",
+        PENDING: "Dochvam DIlmeH, vIpoQ",
+        ACCEPTED: "Do' rur!",
+        REJECTED: "Dochvam luHutlh",
+    },
+};
+export default Klingon;
