+++ packages/core/base/src/utils/i18n/dictionaries/deDE.ts
@@ -1,0 +1,16 @@
+const deDE = {
+    crossmintPayButtonService: {
+        CONNECTING: "Verbindung wird hergestellt...",
+        BUY_WITH_ETH: "Mit ETH kaufen",
+        BUY_WITH_SOL: "Mit SOL kaufen",
+        BUY_WITH_CREDIT_CARD: "Mit Kreditkarte kaufen",
+    },
+    crossmintStatusButtonService: {
+        INVALID: "Ungültige clientId",
+        WAITING_SUBMISSION: "Hier klicken, um Crossmint einzurichten",
+        PENDING: "Ihre Anfrage wird geprüft",
+        ACCEPTED: "Sie sind startklar!",
+        REJECTED: "Ihre Anfrage wurde abgelehnt",
+    },
+};
+export default deDE;
