+++ packages/core/base/src/utils/i18n/dictionaries/enUS.ts
@@ -1,0 +1,16 @@
+const enUS = {
+    crossmintPayButtonService: {
+        CONNECTING: "Connecting...",
+        BUY_WITH_ETH: "Buy with ETH",
+        BUY_WITH_SOL: "Buy with SOL",
+        BUY_WITH_CREDIT_CARD: "Buy with credit card",
+    },
+    crossmintStatusButtonService: {
+        INVALID: "Invalid clientId",
+        WAITING_SUBMISSION: "Click here to setup Crossmint",
+        PENDING: "Your application is under review",
+        ACCEPTED: "You're good to go!",
+        REJECTED: "Your application was rejected",
+    },
+};
+export default enUS;
