+++ packages/core/base/src/utils/i18n/dictionaries/esES.ts
@@ -1,0 +1,16 @@
+const esES = {
+    crossmintPayButtonService: {
+        CONNECTING: "Conectando...",
+        BUY_WITH_ETH: "Comprar con ETH",
+        BUY_WITH_SOL: "Comprar con SOL",
+        BUY_WITH_CREDIT_CARD: "Comprar con tarjeta de crédito",
+    },
+    crossmintStatusButtonService: {
+        INVALID: "clientId inválido",
+        WAITING_SUBMISSION: "Haga clic aquí para configurar Crossmint",
+        PENDING: "Su solicitud está en revisión",
+        ACCEPTED: "¡Está listo para usar Crossmint!",
+        REJECTED: "Su solicitud fue rechazada",
+    },
+};
+export default esES;
