+++ packages/core/base/src/utils/i18n/dictionaries/frFR.ts
@@ -1,0 +1,16 @@
+const frFR = {
+    crossmintPayButtonService: {
+        CONNECTING: "Connexion...",
+        BUY_WITH_ETH: "Acheter avec ETH",
+        BUY_WITH_SOL: "Acheter avec SOL",
+        BUY_WITH_CREDIT_CARD: "Acheter avec une carte de crédit",
+    },
+    crossmintStatusButtonService: {
+        INVALID: "clientId invalide",
+        WAITING_SUBMISSION: "Cliquez ici pour configurer Crossmint",
+        PENDING: "Votre demande est en cours d'examen",
+        ACCEPTED: "Vous êtes prêt à partir !",
+        REJECTED: "Votre demande a été refusée",
+    },
+};
+export default frFR;
