+++ packages/core/base/src/utils/i18n/dictionaries/itIT.ts
@@ -1,0 +1,16 @@
+const itIT = {
+    crossmintPayButtonService: {
+        CONNECTING: "Connessione...",
+        BUY_WITH_ETH: "Acquista con ETH",
+        BUY_WITH_SOL: "Acquista con SOL",
+        BUY_WITH_CREDIT_CARD: "Acquista con carta di credito",
+    },
+    crossmintStatusButtonService: {
+        INVALID: "clientId non valido",
+        WAITING_SUBMISSION: "Clicca qui per configurare Crossmint",
+        PENDING: "La tua richiesta è in fase di revisione",
+        ACCEPTED: "Sei pronto per iniziare!",
+        REJECTED: "La tua richiesta è stata respinta",
+    },
+};
+export default itIT;
