+++ packages/core/base/src/utils/i18n/dictionaries/koKR.ts
@@ -1,0 +1,16 @@
+const koKR = {
+    crossmintPayButtonService: {
+        CONNECTING: "연결 중...",
+        BUY_WITH_ETH: "이더리움으로 구매",
+        BUY_WITH_SOL: "솔라나로 구매",
+        BUY_WITH_CREDIT_CARD: "신용카드로 구매",
+    },
+    crossmintStatusButtonService: {
+        INVALID: "유효하지 않은 clientId",
+        WAITING_SUBMISSION: "여기를 클릭하여 크로스민트 설정",
+        PENDING: "신청이 검토 중입니다",
+        ACCEPTED: "이제 사용할 수 있습니다!",
+        REJECTED: "신청이 거부되었습니다",
+    },
+};
+export default koKR;
