+++ packages/core/base/src/utils/i18n/dictionaries/ptPT.ts
@@ -1,0 +1,16 @@
+const ptPT = {
+    crossmintPayButtonService: {
+        CONNECTING: "A conectar...",
+        BUY_WITH_ETH: "Comprar com ETH",
+        BUY_WITH_SOL: "Comprar com SOL",
+        BUY_WITH_CREDIT_CARD: "Comprar com cartão de crédito",
+    },
+    crossmintStatusButtonService: {
+        INVALID: "clientId inválido",
+        WAITING_SUBMISSION: "Clique aqui para configurar o Crossmint",
+        PENDING: "A sua candidatura está em análise",
+        ACCEPTED: "Está tudo pronto!",
+        REJECTED: "A sua candidatura foi rejeitada",
+    },
+};
+export default ptPT;
