+++ packages/core/base/src/utils/i18n/dictionaries/ruRU.ts
@@ -1,0 +1,16 @@
+const ruRU = {
+    crossmintPayButtonService: {
+        CONNECTING: "Подключение...",
+        BUY_WITH_ETH: "Купить за ETH",
+        BUY_WITH_SOL: "Купить за SOL",
+        BUY_WITH_CREDIT_CARD: "Купить с помощью кредитной карты",
+    },
+    crossmintStatusButtonService: {
+        INVALID: "Недействительный clientId",
+        WAITING_SUBMISSION: "Нажмите здесь, чтобы настроить Crossmint",
+        PENDING: "Ваша заявка на рассмотрении",
+        ACCEPTED: "Вы готовы к работе!",
+        REJECTED: "Ваша заявка была отклонена",
+    },
+};
+export default ruRU;
