+++ packages/core/base/src/utils/i18n/dictionaries/thTH.ts
@@ -1,0 +1,16 @@
+const thTH = {
+    crossmintPayButtonService: {
+        CONNECTING: "กำลังเชื่อมต่อ...",
+        BUY_WITH_ETH: "ซื้อด้วย ETH",
+        BUY_WITH_SOL: "ซื้อด้วย SOL",
+        BUY_WITH_CREDIT_CARD: "ซื้อด้วยบัตรเครดิต",
+    },
+    crossmintStatusButtonService: {
+        INVALID: "clientId ไม่ถูกต้อง",
+        WAITING_SUBMISSION: "คลิกที่นี่เพื่อตั้งค่า Crossmint",
+        PENDING: "ใบสมัครของคุณกำลังอยู่ในระหว่างการตรวจสอบ",
+        ACCEPTED: "คุณพร้อมแล้ว!",
+        REJECTED: "ใบสมัครของคุณถูกปฏิเสธ",
+    },
+};
+export default thTH;
