+++ packages/core/base/src/utils/i18n/dictionaries/trTR.ts
@@ -1,0 +1,16 @@
+const trTR = {
+    crossmintPayButtonService: {
+        CONNECTING: "Bağlanıyor...",
+        BUY_WITH_ETH: "ETH ile Satın Al",
+        BUY_WITH_SOL: "SOL ile Satın Al",
+        BUY_WITH_CREDIT_CARD: "Kredi Kartı ile Satın Al",
+    },
+    crossmintStatusButtonService: {
+        INVALID: "Geçersiz clientId",
+        WAITING_SUBMISSION: "Crossmint'i ayarlamak için buraya tıklayın",
+        PENDING: "Başvurunuz inceleniyor",
+        ACCEPTED: "İşlem yapmaya hazırsınız!",
+        REJECTED: "Başvurunuz reddedildi",
+    },
+};
+export default trTR;
