+++ packages/core/base/src/utils/i18n/dictionaries/ukUA.ts
@@ -1,0 +1,16 @@
+const ukUA = {
+    crossmintPayButtonService: {
+        CONNECTING: "Підключення...",
+        BUY_WITH_ETH: "Купити за ETH",
+        BUY_WITH_SOL: "Купити за SOL",
+        BUY_WITH_CREDIT_CARD: "Купити за допомогою кредитної картки",
+    },
+    crossmintStatusButtonService: {
+        INVALID: "Недійсний clientId",
+        WAITING_SUBMISSION: "Натисніть тут, щоб налаштувати Crossmint",
+        PENDING: "Ваша заявка розглядається",
+        ACCEPTED: "Ви готові до роботи!",
+        REJECTED: "Вашу заявку відхилено",
+    },
+};
+export default ukUA;
