+++ packages/core/base/src/utils/i18n/dictionaries/zhCN.ts
@@ -1,0 +1,16 @@
+const zhCN = {
+    crossmintPayButtonService: {
+        CONNECTING: "连接中...",
+        BUY_WITH_ETH: "使用ETH购买",
+        BUY_WITH_SOL: "使用SOL购买",
+        BUY_WITH_CREDIT_CARD: "使用信用卡购买",
+    },
+    crossmintStatusButtonService: {
+        INVALID: "无效的clientId",
+        WAITING_SUBMISSION: "点击此处设置Crossmint",
+        PENDING: "您的申请正在审查中",
+        ACCEPTED: "您已准备就绪！",
+        REJECTED: "您的申请已被拒绝",
+    },
+};
+export default zhCN;
