+++ packages/core/base/src/utils/i18n/dictionaries/zhTW.ts
@@ -1,0 +1,16 @@
+const zhTW = {
+    crossmintPayButtonService: {
+        CONNECTING: "連接中...",
+        BUY_WITH_ETH: "用ETH購買",
+        BUY_WITH_SOL: "用SOL購買",
+        BUY_WITH_CREDIT_CARD: "用信用卡購買",
+    },
+    crossmintStatusButtonService: {
+        INVALID: "無效的clientId",
+        WAITING_SUBMISSION: "點擊此處設置Crossmint",
+        PENDING: "您的申請正在審查中",
+        ACCEPTED: "您已準備就緒！",
+        REJECTED: "您的申請已被拒絕",
+    },
+};
+export default zhTW;
